new-segment/requirements.txt
matplotlib==3.3.4
numpy==1.24.4
opencv_contrib_python==4.9.0.80
opencv_python==4.5.5.64
opencv_python_headless==4.9.0.80
Pillow==10.2.0
scikit_learn==1.3.2
segmentation_models_pytorch==0.3.3
torch==2.2.1
torchvision==0.17.1
tqdm==4.59.0
